+++ int/rosenbrock_adj.f90
@@ -2523,7 +2523,7 @@ SUBROUTINE FunTemplate( T, Y, Ydot )
    Told = TIME
    TIME = T
    IF ( Do_Update_SUN    ) CALL Update_SUN()
-   IF ( Do_Update_RCONST ) CALL Update_RCONST()
+   IF ( Do_Update_RCONST ) CALL Update_RCONST(Y)
    CALL Fun( Y, FIX, RCONST, Ydot )
    TIME = Told
 
@@ -2554,7 +2554,7 @@ SUBROUTINE JacTemplate( T, Y, Jcb )
     Told = TIME
     TIME = T
     IF ( Do_Update_SUN    ) CALL Update_SUN()
-    IF ( Do_Update_RCONST ) CALL Update_RCONST()
+    IF ( Do_Update_RCONST ) CALL Update_RCONST(Y)
 #ifdef FULL_ALGEBRA
     CALL Jac_SP(Y, FIX, RCONST, JV)
     DO j=1,NVAR
@@ -2591,7 +2591,7 @@ SUBROUTINE HessTemplate( T, Y, Hes )
     Told = TIME
     TIME = T
     IF ( Do_Update_SUN    ) CALL Update_SUN()
-    IF ( Do_Update_RCONST ) CALL Update_RCONST()
+    IF ( Do_Update_RCONST ) CALL Update_RCONST(Y)
     CALL Hessian( Y, FIX, RCONST, Hes )
     TIME = Told